+++ integration/form_utility_functions.go
@@ -1,8 +1,6 @@
 package integration
 
 import (
-	"bytes"
-	"context"
 	"crypto/sha256"
 	"encoding/base64"
 	"encoding/hex"
@@ -10,130 +8,30 @@ import (
 	"math/rand"
 	"strconv"
 	"strings"
-	"sync"
-	"time"
 
 	"github.com/dedis/d-voting/contracts/evoting"
 	"github.com/dedis/d-voting/contracts/evoting/types"
 	"github.com/dedis/d-voting/internal/testing/fake"
 	"github.com/dedis/d-voting/services/dkg"
-	"github.com/dedis/d-voting/services/shuffle"
 	"go.dedis.ch/dela/core/execution/native"
 	"go.dedis.ch/dela/core/ordering"
 	"go.dedis.ch/dela/core/txn"
-	"go.dedis.ch/dela/core/txn/signed"
-	"go.dedis.ch/dela/crypto"
 	"go.dedis.ch/dela/serde"
 	"go.dedis.ch/dela/serde/json"
 	"go.dedis.ch/kyber/v3"
 	"golang.org/x/xerrors"
 )
 
-const addAndWaitErr = "failed to addAndWait: %v"
-
 var serdecontext = json.NewContext()
+
+func encodeID(ID string) types.ID {
+	return types.ID(base64.StdEncoding.EncodeToString([]byte(ID)))
+}
 
 func ballotIsNull(ballot types.Ballot) bool {
 	return ballot.SelectResultIDs == nil && ballot.SelectResult == nil &&
 		ballot.RankResultIDs == nil && ballot.RankResult == nil &&
 		ballot.TextResultIDs == nil && ballot.TextResult == nil
-}
-
-func newTxManager(signer crypto.Signer, firstNode dVotingCosiDela,
-	timeout time.Duration, retry int) txManager {
-
-	client := client{
-		srvc: firstNode.GetOrdering(),
-		mgr:  firstNode.GetValidationSrv(),
-	}
-
-	return txManager{
-		m:     signed.NewManager(signer, client),
-		n:     firstNode,
-		t:     timeout,
-		retry: retry,
-	}
-}
-
-type txManager struct {
-	m     txn.Manager
-	n     dVotingCosiDela
-	t     time.Duration
-	retry int
-}
-
-func (m txManager) addAndWait(args ...txn.Arg) ([]byte, error) {
-	for i := 0; i < m.retry; i++ {
-		sentTxn, err := m.m.Make(args...)
-		if err != nil {
-			return nil, xerrors.Errorf("failed to Make: %v", err)
-		}
-
-		ctx, cancel := context.WithTimeout(context.Background(), m.t)
-		defer cancel()
-
-		events := m.n.GetOrdering().Watch(ctx)
-
-		err = m.n.GetPool().Add(sentTxn)
-		if err != nil {
-			return nil, xerrors.Errorf("failed to Add: %v", err)
-		}
-
-		sentTxnID := sentTxn.GetID()
-
-		accepted := isAccepted(events, sentTxnID)
-		if accepted {
-			return sentTxnID, nil
-		}
-
-		err = m.m.Sync()
-		if err != nil {
-			return nil, xerrors.Errorf("failed to sync: %v", err)
-		}
-
-		cancel()
-	}
-
-	return nil, xerrors.Errorf("transaction not included after timeout: %v", args)
-}
-
-// isAccepted returns true if the transaction was included then accepted
-func isAccepted(events <-chan ordering.Event, txID []byte) bool {
-	for event := range events {
-		for _, result := range event.Transactions {
-			fetchedTxnID := result.GetTransaction().GetID()
-
-			if bytes.Equal(txID, fetchedTxnID) {
-				accepted, _ := event.Transactions[0].GetStatus()
-
-				return accepted
-			}
-		}
-	}
-
-	return false
-}
-
-func grantAccess(m txManager, signer crypto.Signer) error {
-	pubKeyBuf, err := signer.GetPublicKey().MarshalBinary()
-	if err != nil {
-		return xerrors.Errorf("failed to GetPublicKey: %v", err)
-	}
-
-	args := []txn.Arg{
-		{Key: native.ContractArg, Value: []byte("go.dedis.ch/dela.Access")},
-		{Key: "access:grant_id", Value: []byte(hex.EncodeToString(evotingAccessKey[:]))},
-		{Key: "access:grant_contract", Value: []byte("go.dedis.ch/dela.Evoting")},
-		{Key: "access:grant_command", Value: []byte("all")},
-		{Key: "access:identity", Value: []byte(base64.StdEncoding.EncodeToString(pubKeyBuf))},
-		{Key: "access:command", Value: []byte("GRANT")},
-	}
-	_, err = m.addAndWait(args...)
-	if err != nil {
-		return xerrors.Errorf("failed to grantAccess: %v", err)
-	}
-
-	return nil
 }
 
 func createForm(m txManager, title string, admin string) ([]byte, error) {
@@ -382,49 +280,6 @@ func closeForm(m txManager, formID []byte, admin string) error {
 	return nil
 }
 
-func initDkg(nodes []dVotingCosiDela, formID []byte, m txn.Manager) (dkg.Actor, error) {
-	var actor dkg.Actor
-	var err error
-
-	for _, node := range nodes {
-		d := node.(dVotingNode).GetDkg()
-
-		// put Listen in a goroutine to optimize for speed
-		actor, err = d.Listen(formID, m)
-		if err != nil {
-			return nil, xerrors.Errorf("failed to GetDkg: %v", err)
-		}
-	}
-
-	_, err = actor.Setup()
-	if err != nil {
-		return nil, xerrors.Errorf("failed to Setup: %v", err)
-	}
-
-	return actor, nil
-}
-
-func initShuffle(nodes []dVotingCosiDela) (shuffle.Actor, error) {
-	var sActor shuffle.Actor
-
-	for _, node := range nodes {
-		client := client{
-			srvc: node.GetOrdering(),
-			mgr:  node.GetValidationSrv(),
-		}
-
-		var err error
-		shuffler := node.GetShuffle()
-
-		sActor, err = shuffler.Listen(signed.NewManager(node.GetShuffleSigner(), client))
-		if err != nil {
-			return nil, xerrors.Errorf("failed to init Shuffle: %v", err)
-		}
-	}
-
-	return sActor, nil
-}
-
 func decryptBallots(m txManager, actor dkg.Actor, form types.Form) error {
 	if form.Status != types.PubSharesSubmitted {
 		return xerrors.Errorf("cannot decrypt: not all pubShares submitted")
@@ -451,74 +306,4 @@ func decryptBallots(m txManager, actor dkg.Actor, form types.Form) error {
 	}
 
 	return nil
-}
-
-func closeNodes(nodes []dVotingCosiDela) error {
-	wait := sync.WaitGroup{}
-	wait.Add(len(nodes))
-
-	for _, n := range nodes {
-		go func(node dVotingNode) {
-			defer wait.Done()
-			node.GetOrdering().Close()
-		}(n.(dVotingNode))
-	}
-
-	done := make(chan struct{})
-
-	go func() {
-		wait.Wait()
-		close(done)
-	}()
-
-	select {
-	case <-done:
-		return nil
-	case <-time.After(time.Second * 30):
-		return xerrors.New("failed to close: timeout")
-	}
-}
-
-func encodeID(ID string) types.ID {
-	return types.ID(base64.StdEncoding.EncodeToString([]byte(ID)))
-}
-
-// waitForStatus polls the nodes until they all updated to the expected status
-// for the given form. An error is raised if the timeout expires.
-func waitForStatus(status types.Status, formFac types.FormFactory,
-	formID []byte, nodes []dVotingCosiDela, numNodes int, timeOut time.Duration) error {
-
-	expiration := time.Now().Add(timeOut)
-
-	isOK := func() (bool, error) {
-		for _, node := range nodes {
-			form, err := getForm(formFac, formID, node.GetOrdering())
-			if err != nil {
-				return false, xerrors.Errorf("failed to get form: %v", err)
-			}
-
-			if form.Status != status {
-				return false, nil
-			}
-		}
-
-		return true, nil
-	}
-
-	for {
-		if time.Now().After(expiration) {
-			return xerrors.New("status check expired")
-		}
-
-		ok, err := isOK()
-		if err != nil {
-			return xerrors.Errorf("failed to check status: %v", err)
-		}
-
-		if ok {
-			return nil
-		}
-
-		time.Sleep(time.Millisecond * 100)
-	}
 }
